% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackage.R
\docType{package}
\name{CreatePackageInR-package}
\alias{CreatePackageInR}
\alias{CreatePackageInR-package}
\title{CreatePackageInR: A Package for Demonstrating Custom Functions in R}
\description{
The CreatePackageInR package provides a collection of functions that are used for
graph creation and mathematical calculations, such as finding the greatest common divisor (GCD)
and calculating shortest paths using Dijkstra's algorithm.
}
\section{CreatePackageInR functions}{

The package includes the following functions:
\itemize{
  \item \code{euclidean()}: Implements the Euclidean algorithm to find the GCD of two numbers.
  \item \code{dijkstra()}: Implements Dijkstra's algorithm to find the shortest path from a given initial node.
}
}

\keyword{internal}
