% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackage.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm to find GCDs}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{First number}

\item{b}{Second number}
}
\value{
GCD of a and b
}
\description{
Euclidean Algorithm to find GCDs
}
\examples{
euclidean(100, 1000)
euclidean(123612, 13892347912)
}
